+++ CHANGELOG.md
@@ -31,6 +31,8 @@
 - Navigation pane toggle becoming unresponsive if the navigation pane was closed with the toggle on touch screen
 - Some HLS-type radio streams not playing because URLs contained in `#EXT-X-*` headers were handled incorrectly
   [#133](https://github.com/nc-music/music/issues/133)
+- Scrobbling to Last.fm failing if the song was missing either track number or album name
+  [#134](https://github.com/nc-music/music/issues/134)
 - Subsonic API:
   * Response of `getArtistInfo` and `getArtistInfo2` being malformed when no details available
     [#129](https://github.com/nc-music/music/issues/129)